# src/apps/backend-spring/src/main/java/org/openapitools/api/PlaygroundsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Playground;
import org.openapitools.model.PlaygroundsGet200Response;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-07-29T23:35:34.193120+02:00[Europe/Berlin]", comments = "Generator version: 7.14.0")
@Validated
@Tag(name = "Playgrounds", description = "the Playgrounds API")
public interface PlaygroundsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /playgrounds : Get all playgrounds
     *
     * @param limit Max number of playgrounds to return (optional)
     * @param skip Number of playgrounds to skip (optional)
     * @return List of playgrounds (status code 200)
     */
    @Operation(
        operationId = "playgroundsGet",
        summary = "Get all playgrounds",
        tags = { "Playgrounds" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of playgrounds", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PlaygroundsGet200Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/playgrounds",
        produces = { "application/json" }
    )
    
    default ResponseEntity<PlaygroundsGet200Response> playgroundsGet(
        @Parameter(name = "limit", description = "Max number of playgrounds to return", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false) @Nullable Integer limit,
        @Parameter(name = "skip", description = "Number of playgrounds to skip", in = ParameterIn.QUERY) @Valid @RequestParam(value = "skip", required = false) @Nullable Integer skip
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"total\" : 6, \"limit\" : 5, \"skip\" : 1, \"playgrounds\" : [ { \"PlaygroundID\" : 0, \"Name\" : \"Name\", \"Location\" : \"Location\" }, { \"PlaygroundID\" : 0, \"Name\" : \"Name\", \"Location\" : \"Location\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /playgrounds/{playgroundId} : Get playground by ID
     *
     * @param playgroundId  (required)
     * @return Playground details (status code 200)
     *         or Playground not found (status code 404)
     */
    @Operation(
        operationId = "playgroundsPlaygroundIdGet",
        summary = "Get playground by ID",
        tags = { "Playgrounds" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Playground details", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Playground.class))
            }),
            @ApiResponse(responseCode = "404", description = "Playground not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/playgrounds/{playgroundId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Playground> playgroundsPlaygroundIdGet(
        @Parameter(name = "playgroundId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("playgroundId") Integer playgroundId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"PlaygroundID\" : 0, \"Name\" : \"Name\", \"Location\" : \"Location\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}